requirements-sleep.txt
pandas
numpy
python-bson
# If you use Git LFS for large CSVs: git-lfs
